requirements.txt
editdistance~=0.6.0
lmdb~=1.3.0
matplotlib
numpy==1.23.1
opencv-python~=4.6.0.66
path~=16.4.0
# tensorflow
flask~=2.2.1
Image
pillow~=9.2.0
werkzeug~=2.2.1